Fix dom store
